add client for testing

+++ 6/message/dispatcher_test.go
@@ -16,13 +16,13 @@ func TestIAmDispatcher(t *testing.T) {
 		{
 			Msg: []byte{0x81, 0x01, 0x00, 0x42},
 			Expected: &IAmDispatcher{
-				Roads: []int{66},
+				Roads: []uint16{66},
 			},
 		},
 		{
 			Msg: []byte{0x81, 0x03, 0x00, 0x42, 0x01, 0x70, 0x13, 0x88},
 			Expected: &IAmDispatcher{
-				Roads: []int{66, 368, 5000},
+				Roads: []uint16{66, 368, 5000},
 			},
 		},
 	}
